requirements.txt
anyio==3.5.0
asgiref==3.4.1
autopep8==1.6.0
chardet==4.0.0
click==8.0.3
Faker==11.3.0
fastapi==0.72.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
idna==3.3
numpy==1.22.1
pandas==1.3.5
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pydantic==1.9.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.30
starlette==0.17.1
text-unidecode==1.3
toml==0.10.2
typing_extensions==4.0.1
uvicorn==0.17.0
